requirements.txt
streamlit==1.29.0
openai==1.12.0
pillow==10.1.0
numpy==1.26.2
scipy==1.11.4
sympy==1.12
matplotlib==3.8.2
requests==2.31.0
python-dotenv==1.0.0
scikit-image==0.22.0
tensorflow-hub==0.14.0
langchain==0.0.335
pydantic==2.5.2
pytest==7.4.3
huggingface_hub==0.19.4
httpx==0.25.0 